refactor(automation): extract repo-management cluster into loop_repo_manager.py (#1360)
Extract 12 pure-function repo-management helpers from loop_runner.py
(1,721 lines) into a new module loop_repo_manager.py (~403 lines),
reducing loop_runner.py by ~22% to ~1,370 lines.

Functions extracted:
- _detect_cwd_repo, _gh_list_repos, _list_open_issue_numbers
- _count_open_issues, _count_failing_prs, _sort_repos_by_open_count
- _resolve_repo_dir, _ensure_clone, _clone_missing_repos
- _detect_remote_base_ref, _local_ahead_count, _rebase_main

Backward-compat re-exports added to loop_runner.py so all existing call
sites and the majority of test mock paths keep resolving unchanged.

Test patches targeting loop_repo_manager internals (resilient_call, LOG)
updated to the new module path per patch-where-looked-up convention.

Also adds loop_repo_manager.py to the frozen coverage omit allowlist
(17 modules, up from 16) and unit tests in test_loop_repo_manager.py.

Part of umbrella #1179.

Closes #1360
